move float and github related classes out of core
