+++ float/src/main/java/com/novoda/floatschedule/convert/FloatGithubProjectConverter.java
@@ -1,6 +1,6 @@
 package com.novoda.floatschedule.convert;
 
-import com.novoda.github.reports.reader.ProjectsReader;
+import com.novoda.floatschedule.reader.ProjectsReader;
 
 import java.io.IOException;
 import java.util.List;
