+++ float/src/main/java/com/novoda/floatschedule/project/ProjectsServiceClient.java
@@ -1,4 +1,6 @@
-package com.novoda.github.reports.reader;
+package com.novoda.floatschedule.project;
+
+import com.novoda.floatschedule.reader.ProjectsReader;
 
 import java.io.IOException;
 import java.util.List;
@@ -7,8 +9,6 @@ import java.util.stream.Collectors;
 import rx.Observable;
 
 public class ProjectsServiceClient {
-
-    // FIXME we're leaking float into core
 
     private final ProjectsReader projectsReader;
 
@@ -20,7 +20,7 @@ public class ProjectsServiceClient {
         this.projectsReader = projectsReader;
     }
 
-    Observable<String> getAllProjectNames() {
+    public Observable<String> getAllProjectNames() {
         try {
             readIfNeeded();
             return Observable.from(projectsReader.getContent().keySet());
@@ -30,7 +30,7 @@ public class ProjectsServiceClient {
         return Observable.empty();
     }
 
-    Observable<String> getAllGithubRepositoryNames(List<String> projectNames) {
+    public Observable<String> getAllGithubRepositoryNames(List<String> projectNames) {
         try {
             readIfNeeded();
