+++ float/src/main/java/com/novoda/floatschedule/reader/ProjectsReader.java
@@ -1,4 +1,4 @@
-package com.novoda.github.reports.reader;
+package com.novoda.floatschedule.reader;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
@@ -7,8 +7,6 @@ import java.util.List;
 import java.util.Map;
 
 public class ProjectsReader {
-
-    // FIXME we're leaking float into core
 
     private final JsonMapReader<Map<String, List<String>>> jsonMapReader;
     private final Map<String, List<String>> projectToRepositories;
