+++ float/src/main/java/com/novoda/floatschedule/reader/UsersReader.java
@@ -1,4 +1,4 @@
-package com.novoda.github.reports.reader;
+package com.novoda.floatschedule.reader;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
@@ -6,8 +6,6 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class UsersReader {
-
-    // FIXME we're leaking float into core
 
     private final JsonMapReader<Map<String, String>> jsonMapReader;
     private final Map<String, String> floatToGithubUser;
