+++ float/src/test/resources/stringToString.json
@@ -1,0 +1,5 @@
+{
+  "key0": "value0",
+  "key1": "value1",
+  "key2": "value2"
+}
