+++ float/src/test/resources/stringToStrings.json
@@ -1,0 +1,5 @@
+{
+  "key0": [ "value0a", "value0b" ],
+  "key1": [ "value1" ],
+  "key2": [ "value2a", "value2b", "value2c" ]
+}
