+++ web-service/src/main/java/com/novoda/github/reports/web/lambda/UsersServiceClient.java
@@ -1,12 +1,12 @@
-package com.novoda.github.reports.reader;
+package com.novoda.github.reports.web.lambda;
+
+import com.novoda.floatschedule.reader.UsersReader;
 
 import java.io.IOException;
 
 import rx.Observable;
 
 public class UsersServiceClient {
-
-    // FIXME we're leaking float into core
 
     private final UsersReader usersReader;
